# Some FermionicOp speed-improvements