### Summary

This PR introduces some small improvements to the `FermionicOp` in order to speedup the multiplication of two operators.


### Details and comments

In a discussion with @1ucian0 we identified some inefficiencies in the `FermionicOp._single_mul` method. This PR removes these inefficiencies and achieves a speedup of 2x in some scenarios.

Old code:
```python
python -m timeit -n 10000 -s "from qiskit_nature.operators.second_quantization import FermionicOp" "FermionicOp('I'*100) @ FermionicOp('I'*100)"
10000 loops, best of 5: 51.2 usec per loop
python -m timeit -n 10000 -s "from qiskit_nature.operators.second_quantization import FermionicOp" "FermionicOp('I'*100) @ FermionicOp('+'*100)"
10000 loops, best of 5: 97.2 usec per loop
python -m timeit -n 10000 -s "from qiskit_nature.operators.second_quantization import FermionicOp" "FermionicOp('+'*100) @ FermionicOp('+'*100)"
10000 loops, best of 5: 22 usec per loop
python -m timeit -n 10000 -s "from qiskit_nature.operators.second_quantization import FermionicOp" "FermionicOp('+'*100) @ FermionicOp('-'*100)"
10000 loops, best of 5: 116 usec per loop
python -m timeit -n 10000 -s "from qiskit_nature.operators.second_quantization import FermionicOp" "FermionicOp('+'*100) @ FermionicOp('N'*100)"
10000 loops, best of 5: 23.8 usec per loop
python -m timeit -n 10000 -s "from qiskit_nature.operators.second_quantization import FermionicOp" "FermionicOp('N'*100) @ FermionicOp('+'*100)"
10000 loops, best of 5: 98.2 usec per loop
python -m timeit -n 10000 -s "from qiskit_nature.operators.second_quantization import FermionicOp" "FermionicOp('N'*100) @ FermionicOp('N'*100)"
10000 loops, best of 5: 62.3 usec per loop
```

New code:
```python
python -m timeit -n 10000 -s "from qiskit_nature.operators.second_quantization import FermionicOp" "FermionicOp('I'*100) @ FermionicOp('I'*100)"
10000 loops, best of 5: 36.7 usec per loop
python -m timeit -n 10000 -s "from qiskit_nature.operators.second_quantization import FermionicOp" "FermionicOp('I'*100) @ FermionicOp('+'*100)"
10000 loops, best of 5: 40 usec per loop
python -m timeit -n 10000 -s "from qiskit_nature.operators.second_quantization import FermionicOp" "FermionicOp('+'*100) @ FermionicOp('+'*100)"
10000 loops, best of 5: 19.7 usec per loop
python -m timeit -n 10000 -s "from qiskit_nature.operators.second_quantization import FermionicOp" "FermionicOp('+'*100) @ FermionicOp('-'*100)"
10000 loops, best of 5: 55.6 usec per loop
python -m timeit -n 10000 -s "from qiskit_nature.operators.second_quantization import FermionicOp" "FermionicOp('+'*100) @ FermionicOp('N'*100)"
10000 loops, best of 5: 24.2 usec per loop
python -m timeit -n 10000 -s "from qiskit_nature.operators.second_quantization import FermionicOp" "FermionicOp('N'*100) @ FermionicOp('+'*100)"
10000 loops, best of 5: 48.2 usec per loop
python -m timeit -n 10000 -s "from qiskit_nature.operators.second_quantization import FermionicOp" "FermionicOp('N'*100) @ FermionicOp('N'*100)"
10000 loops, best of 5: 56 usec per loop
```