@mrossinek 2021-02-08
The remaining CI failures are related to an error in Aqua. Fix pending: https://github.com/Qiskit/qiskit-aqua/pull/1527